Fix different places same item bug
The string in the last searches CSV file now contains also the location
so that the same item in different places is now written in a separate
line.
Fix the link with no region that was not working.
Change populate_search method to add a new field called search_description
Change search_item to handle search_description field
Add search_description in Fields class.